refactor(jsruntime): optimize the size of generated code
